+++ src/gamesvr/mainlogic/activity_open_fund_msg.go
@@ -47,7 +47,7 @@ func Hand_GetOpenFundStatus(w http.ResponseWriter, r *http.Request) {
 	response.CostMoneyNum = gamedata.OpenFundPriceNum
 
 	for _, v := range gamedata.GT_OpenFundLst[gamedata.OpenFund_Level] {
-		if player.ActivityModule.OpenFund.FundLevelMark.Get(uint32(v.ID)) != true {
+		if player.ActivityModule.OpenFund.FundLevelMark.Get(v.ID) != true {
 			awardLst := gamedata.GetItemsFromAwardID(v.Award)
 			for _, m := range awardLst {
 				response.ReceiveMoney += m.ItemNum
@@ -115,7 +115,7 @@ func Hand_BuyOpenFund(w http.ResponseWriter, r *http.Request) {
 	response.FundLevelMark = int(player.ActivityModule.OpenFund.FundLevelMark)
 
 	for _, v := range gamedata.GT_OpenFundLst[gamedata.OpenFund_Level] {
-		if player.ActivityModule.OpenFund.FundLevelMark.Get(uint32(v.ID)) != true {
+		if player.ActivityModule.OpenFund.FundLevelMark.Get(v.ID) != true {
 			awardLst := gamedata.GetItemsFromAwardID(v.Award)
 			for _, m := range awardLst {
 				response.ReceiveMoney += m.ItemNum
@@ -178,7 +178,7 @@ func Hand_GetOpenFundAllAward(w http.ResponseWriter, r *http.Request) {
 	}
 
 	//! 判断是否已经领取
-	if player.ActivityModule.OpenFund.FundCountMark.Get(uint32(req.ID)) == true {
+	if player.ActivityModule.OpenFund.FundCountMark.Get(req.ID) == true {
 		response.RetCode = msg.RE_ALREADY_RECEIVED
 		return
 	}
@@ -193,7 +193,7 @@ func Hand_GetOpenFundAllAward(w http.ResponseWriter, r *http.Request) {
 	}
 
 	//! 改变标记
-	player.ActivityModule.OpenFund.FundCountMark.Set(uint32(req.ID))
+	player.ActivityModule.OpenFund.FundCountMark.Set(req.ID)
 	player.ActivityModule.OpenFund.UpdateFundCountMark()
 
 	//! 返回成功
@@ -253,7 +253,7 @@ func Hand_GetOpenFundLevelAward(w http.ResponseWriter, r *http.Request) {
 	}
 
 	//! 判断是否已经领取
-	if player.ActivityModule.OpenFund.FundLevelMark.Get(uint32(req.ID)) == true {
+	if player.ActivityModule.OpenFund.FundLevelMark.Get(req.ID) == true {
 		response.RetCode = msg.RE_ALREADY_RECEIVED
 		return
 	}
@@ -268,7 +268,7 @@ func Hand_GetOpenFundLevelAward(w http.ResponseWriter, r *http.Request) {
 	}
 
 	//! 改变标记
-	player.ActivityModule.OpenFund.FundLevelMark.Set(uint32(req.ID))
+	player.ActivityModule.OpenFund.FundLevelMark.Set(req.ID)
 	player.ActivityModule.OpenFund.UpdateFundLevelMark()
 
 	//! 返回成功
